requirements.txt
# For hogwild
accelerate==1.6.0
datasets==3.5.0
numpy>=2.2.4
tokenizers==0.21.1
torch>=2.6.0
transformers==4.51.0
wandb>=0.17.2
tqdm

# For eval
latex2mathml==3.78.1

# For tortoise-tts
deepspeed>=0.18.2
unidecode>=1.4.0
inflect>=7.5.0
librosa>=0.11.0
rotary_embedding_torch
einops>=0.8.1
progressbar>=2.5
torchaudio>=2.8.0
setuptools>=80.9.0
rpyc==6.0.1

# For voice assistant demo
ffmpeg-python==0.2.0
openai-whisper==20250625
Jinja2==3.1.6
